repeat系のマクロの添字をlong longに変更

+++ template.cpp
@@ -16,10 +16,10 @@ using namespace std;
 typedef long long ll;
 typedef pair<int, int> p_i;
 
-#define rep(i, n) for(int i = 0; i < n; i++)
-#define repext(i, a, b) for (int i = a; i < b; i++)
-#define reprev(i, n) for (int i = n - 1; i >= 0; i--)
-#define repextrev(i, a, b) for (int i = b - 1; i >= a; i--)
+#define rep(i, n) for(ll i = 0LL; i < n; i++)
+#define repext(i, a, b) for (ll i = a; i < b; i++)
+#define reprev(i, n) for (ll i = n - 1; i >= 0; i--)
+#define repextrev(i, a, b) for (ll i = b - 1; i >= a; i--)
 
 #define repv(v, values) for (auto v : values)
